% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dftomat.R
\name{dftomat}
\alias{dftomat}
\title{Convert Data Frame to Numeric Matrix}
\usage{
dftomat(X)
}
\arguments{
\item{X}{A data frame to be converted to a numeric matrix.}
}
\value{
A numeric matrix corresponding to the input data frame.
}
\description{
This function converts a data frame into a numeric matrix. If the data frame
contains factor or character variables, they are first converted to numeric.
}
\keyword{internal}
